requirements.txt
# Python 3.12 Compatible Requirements
# Updated versions for Python 3.12 compatibility

# Image Processing & Math
numpy>=1.26.0
Pillow>=10.0.0
scipy>=1.11.0
PyWavelets>=1.4.1
ImageHash>=4.3.1

# HTTP and Networking
requests>=2.31.0
urllib3>=2.0.3

# Utilities
python-dateutil>=2.8.2
typing_extensions>=4.12.2

# Image Structure Similarity (SSIM)
scikit-image>=0.22.0
